Modularize skill loading internals
Refactor sdk/context/skills/skill.ts into focused modules for discovery, parsing, and validation while preserving behavior.\n\nBead: oh-tab-zhwi.8.2
